chore: 백엔드 api 변경사항 프론트 반영 및 응답 수정
